added presentation order + new song preview
